+++ Sources/CustomerGlu/Classes/Cells/CGClientTestingNudgeSetupView.xib
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="21507" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_12" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21505"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" id="iN0-l3-epB" customClass="CGClientTestingNudgeSetupView" customModule="CustomerGlu">
+            <rect key="frame" x="0.0" y="0.0" width="375" height="36"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="hfF-lS-rMB">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="36"/>
+                    <subviews>
+                        <activityIndicatorView opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" hidesWhenStopped="YES" animating="YES" style="medium" translatesAutoresizingMaskIntoConstraints="NO" id="zma-mQ-g4r">
+                            <rect key="frame" x="0.0" y="8" width="20" height="20"/>
+                            <constraints>
+                                <constraint firstAttribute="height" constant="20" id="G0k-SQ-f6F"/>
+                                <constraint firstAttribute="width" constant="20" id="gZG-SJ-irH"/>
+                            </constraints>
+                        </activityIndicatorView>
+                        <imageView hidden="YES" clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="GreenTick" translatesAutoresizingMaskIntoConstraints="NO" id="0UR-cI-xUT">
+                            <rect key="frame" x="0.0" y="8" width="20" height="20"/>
+                            <constraints>
+                                <constraint firstAttribute="height" constant="20" id="2fW-f3-Jw7"/>
+                                <constraint firstAttribute="width" constant="20" id="Sja-BT-bWg"/>
+                            </constraints>
+                        </imageView>
+                        <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Label" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="gXY-M3-GrF">
+                            <rect key="frame" x="32" y="11" width="327" height="14.333333333333336"/>
+                            <fontDescription key="fontDescription" type="system" pointSize="12"/>
+                            <nil key="textColor"/>
+                            <nil key="highlightedColor"/>
+                        </label>
+                    </subviews>
+                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    <constraints>
+                        <constraint firstItem="zma-mQ-g4r" firstAttribute="leading" secondItem="hfF-lS-rMB" secondAttribute="leading" id="Ar3-iX-0mr"/>
+                        <constraint firstItem="0UR-cI-xUT" firstAttribute="height" secondItem="zma-mQ-g4r" secondAttribute="height" id="CkW-xd-9Ws"/>
+                        <constraint firstAttribute="bottom" secondItem="zma-mQ-g4r" secondAttribute="bottom" constant="8" id="FKC-PK-eZ3"/>
+                        <constraint firstItem="gXY-M3-GrF" firstAttribute="centerY" secondItem="0UR-cI-xUT" secondAttribute="centerY" id="OUE-mV-L4b"/>
+                        <constraint firstItem="gXY-M3-GrF" firstAttribute="leading" secondItem="hfF-lS-rMB" secondAttribute="leading" constant="32" id="PmI-2u-wQz"/>
+                        <constraint firstAttribute="trailing" secondItem="gXY-M3-GrF" secondAttribute="trailing" constant="16" id="era-6N-IAx"/>
+                        <constraint firstItem="0UR-cI-xUT" firstAttribute="centerY" secondItem="zma-mQ-g4r" secondAttribute="centerY" id="lOc-YI-zqF"/>
+                        <constraint firstItem="0UR-cI-xUT" firstAttribute="centerX" secondItem="zma-mQ-g4r" secondAttribute="centerX" id="lc5-R5-UGB"/>
+                        <constraint firstItem="zma-mQ-g4r" firstAttribute="top" secondItem="hfF-lS-rMB" secondAttribute="top" constant="8" id="rcQ-0x-rw1"/>
+                        <constraint firstItem="0UR-cI-xUT" firstAttribute="width" secondItem="zma-mQ-g4r" secondAttribute="width" id="zAt-at-v9P"/>
+                    </constraints>
+                </view>
+            </subviews>
+            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <constraints>
+                <constraint firstAttribute="bottom" secondItem="hfF-lS-rMB" secondAttribute="bottom" id="DYV-4R-zdg"/>
+                <constraint firstAttribute="trailing" secondItem="hfF-lS-rMB" secondAttribute="trailing" id="JPR-Lk-BC8"/>
+                <constraint firstAttribute="top" secondItem="hfF-lS-rMB" secondAttribute="top" id="UMT-Hz-c29"/>
+                <constraint firstItem="hfF-lS-rMB" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="pIC-1q-pdF"/>
+            </constraints>
+            <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
+            <connections>
+                <outlet property="containerView" destination="hfF-lS-rMB" id="w7s-rT-T7I"/>
+                <outlet property="indicatorView" destination="zma-mQ-g4r" id="PdW-Ne-8gz"/>
+                <outlet property="tickImageView" destination="0UR-cI-xUT" id="r33-fi-0hH"/>
+                <outlet property="titleLabel" destination="gXY-M3-GrF" id="eGD-ci-iKr"/>
+                <outlet property="titleLabelLeadingConstraint" destination="PmI-2u-wQz" id="W1h-iE-lYo"/>
+            </connections>
+            <point key="canvasLocation" x="389" y="19"/>
+        </view>
+    </objects>
+    <resources>
+        <image name="GreenTick" width="42.666667938232422" height="42.666667938232422"/>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
