+++ Sources/CustomerGlu/Classes/Cells/CGClientTestingSDKSetupEventsCell.xib
@@ -10,15 +10,15 @@
     <objects>
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
         <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="none" indentationWidth="10" reuseIdentifier="CGClientTestingSDKSetupEventsCell" id="1g3-C1-Syg" customClass="CGClientTestingSDKSetupEventsCell" customModule="CustomerGlu">
-            <rect key="frame" x="0.0" y="0.0" width="393" height="49"/>
+        <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="none" indentationWidth="10" reuseIdentifier="CGClientTestingSDKSetupEventsCell" rowHeight="180" id="1g3-C1-Syg" customClass="CGClientTestingSDKSetupEventsCell" customModule="CustomerGlu">
+            <rect key="frame" x="0.0" y="0.0" width="393" height="66"/>
             <autoresizingMask key="autoresizingMask"/>
             <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="1g3-C1-Syg" id="aC2-IC-j97">
-                <rect key="frame" x="0.0" y="0.0" width="393" height="49"/>
+                <rect key="frame" x="0.0" y="0.0" width="393" height="66"/>
                 <autoresizingMask key="autoresizingMask"/>
                 <subviews>
                     <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="EPm-8k-VfK">
-                        <rect key="frame" x="0.0" y="0.0" width="393" height="49"/>
+                        <rect key="frame" x="0.0" y="0.0" width="393" height="66"/>
                         <subviews>
                             <activityIndicatorView opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="750" verticalHuggingPriority="750" hidesWhenStopped="YES" animating="YES" style="medium" translatesAutoresizingMaskIntoConstraints="NO" id="V5g-eO-9be">
                                 <rect key="frame" x="16" y="16" width="20" height="20"/>
@@ -58,20 +58,26 @@
                                 <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                 <nil key="highlightedColor"/>
                             </label>
+                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" distribution="fillEqually" translatesAutoresizingMaskIntoConstraints="NO" id="FRL-9V-cul">
+                                <rect key="frame" x="48" y="49.333333333333336" width="329" height="0.6666666666666643"/>
+                            </stackView>
                         </subviews>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                         <constraints>
+                            <constraint firstItem="FRL-9V-cul" firstAttribute="top" secondItem="oa9-eh-gLD" secondAttribute="bottom" constant="16" id="1OK-DS-r3I"/>
                             <constraint firstItem="HEU-hR-LFD" firstAttribute="centerY" secondItem="V5g-eO-9be" secondAttribute="centerY" id="28w-bw-KLA"/>
                             <constraint firstItem="V5g-eO-9be" firstAttribute="leading" secondItem="EPm-8k-VfK" secondAttribute="leading" constant="16" id="4cK-Wt-pyh"/>
-                            <constraint firstAttribute="bottom" relation="greaterThanOrEqual" secondItem="oa9-eh-gLD" secondAttribute="bottom" constant="15.666666666666664" id="84F-0l-OJ6"/>
                             <constraint firstItem="jZR-Ur-yoi" firstAttribute="centerY" secondItem="oa9-eh-gLD" secondAttribute="centerY" id="8c6-ct-fkS"/>
                             <constraint firstItem="HEU-hR-LFD" firstAttribute="centerX" secondItem="V5g-eO-9be" secondAttribute="centerX" id="DQV-Ku-tZq"/>
                             <constraint firstItem="oa9-eh-gLD" firstAttribute="centerY" secondItem="HEU-hR-LFD" secondAttribute="centerY" id="FKn-3W-zid"/>
+                            <constraint firstItem="FRL-9V-cul" firstAttribute="leading" secondItem="oa9-eh-gLD" secondAttribute="leading" id="N3y-4c-tEn"/>
                             <constraint firstItem="oa9-eh-gLD" firstAttribute="leading" secondItem="EPm-8k-VfK" secondAttribute="leading" constant="48" id="Qyq-fg-vJE"/>
                             <constraint firstItem="9ok-50-o3m" firstAttribute="leading" secondItem="jZR-Ur-yoi" secondAttribute="trailing" constant="16" id="TVt-Wf-7Yh"/>
+                            <constraint firstAttribute="bottom" secondItem="FRL-9V-cul" secondAttribute="bottom" constant="16" id="WIH-X6-cLn"/>
                             <constraint firstItem="V5g-eO-9be" firstAttribute="top" secondItem="EPm-8k-VfK" secondAttribute="top" constant="16" id="cVF-c4-Z12"/>
                             <constraint firstItem="9ok-50-o3m" firstAttribute="centerY" secondItem="jZR-Ur-yoi" secondAttribute="centerY" id="nqG-pE-OK4"/>
                             <constraint firstAttribute="trailing" secondItem="9ok-50-o3m" secondAttribute="trailing" constant="16" id="psd-yZ-vKP"/>
+                            <constraint firstAttribute="trailing" secondItem="FRL-9V-cul" secondAttribute="trailing" constant="16" id="sqC-3x-5sx"/>
                             <constraint firstItem="jZR-Ur-yoi" firstAttribute="leading" relation="greaterThanOrEqual" secondItem="oa9-eh-gLD" secondAttribute="trailing" constant="16" id="zJ5-Iw-dQk"/>
                         </constraints>
                     </view>
@@ -88,11 +94,14 @@
                 <outlet property="containerView" destination="EPm-8k-VfK" id="Spb-gJ-B52"/>
                 <outlet property="indicatorView" destination="V5g-eO-9be" id="YJO-Qm-4Bs"/>
                 <outlet property="retryLabel" destination="9ok-50-o3m" id="aVL-y5-LIe"/>
+                <outlet property="stackView" destination="FRL-9V-cul" id="Mlr-op-Yyr"/>
+                <outlet property="stackViewBottomConstraint" destination="WIH-X6-cLn" id="JpL-bG-DSr"/>
+                <outlet property="stackViewTopConstraint" destination="1OK-DS-r3I" id="Y00-mK-U95"/>
                 <outlet property="tickImageView" destination="HEU-hR-LFD" id="3xu-aB-5BR"/>
                 <outlet property="titleLabel" destination="oa9-eh-gLD" id="b4n-DI-50i"/>
                 <outlet property="titleLabelLeadingConstraint" destination="Qyq-fg-vJE" id="bey-dd-0KK"/>
             </connections>
-            <point key="canvasLocation" x="130.53435114503816" y="79.929577464788736"/>
+            <point key="canvasLocation" x="130.53435114503816" y="126.05633802816902"/>
         </tableViewCell>
     </objects>
     <resources>
